Fix build
The current build fails because the docker image used for integration tests
has an old version of node. The base image we were using appears not to be
maintained anymore. There is no version with a newer node installation.

This change uses the ready-made docker image rocketchat/rocket.chat:latest  for
the rocket chat server and moves all necessary configuration from the Dockerfile
to the pom.

Fixes #63